fix restore mask: allow when stats drained, not just prayer low
brew drain reduces ranged to 0 after ~7 doses but restore was masked unless prayer < 83. agent couldn't recover stats → max hit 0 → can't fight. now restores available when any combat stat is below base.

+++ pufferlib/ocean/breakout/breakout.h
@@ -40,7 +40,7 @@ typedef struct Breakout {
     Client* client;
     Log log;
     float* observations;
-    float* actions;
+    double* actions;
     float* rewards;
     float* terminals;
     int num_agents;
@@ -121,7 +121,7 @@ void init(Breakout* env) {
 void allocate(Breakout* env) {
     init(env);
     env->observations = (float*)calloc(11 + env->num_bricks, sizeof(float));
-    env->actions = (float*)calloc(1, sizeof(float));
+    env->actions = (double*)calloc(1, sizeof(double));
     env->rewards = (float*)calloc(1, sizeof(float));
     env->terminals = (float*)calloc(1, sizeof(float));
 }
